[PPML]fix pyspark test script (#7891)
* [PPML]add prompt in pyspark sqlApi script

* change success condition

* add PYSPARK_PYTHON

* fix

* fix

* increase epc size of k8s test

* delete MALLOC_ARENA_MAX in driver.yaml

* fix script simpleexample

* decrease epc size in weekly test

* fix sqlApi script

* move the harder ones to the front

* fix

* fix

* fix

* fix
